Improve RegEx pattern for "IF in SELECT statement"
(Full explanation will be in the PR)

RegEx pattern for "IF in SELECT statement" is overly broad in several areas, allowing for both unexpected captures and capturing too much.
I simplified the RegEx pattern to restrict it to:
1. capturing only IF statements (not something like "datedIFf")
2. capturing just the IF statement (not something like "IF (a, b, c) AS alias, d, function(e) as f")
3. capturing multiple IF statements in the query

Also, SQL Server 2012 (the lowest version supported by Project Nami) introduced the "IIF()" function, which is identical to MySQL's "IF()" function, so rather than chopping up the IF function to reconstitute as a CASE WHEN, simply prefix the "IF()" function with an extra "I". Easy peasy. 😸